Moved logic into separate files
- Now each piece will have a dedicated class
- Start of rework of many systems
- MoveGenerationUtils needs a rework
- Utils need a good solution
