+++ src/main/java/DataTypes/Move.java
@@ -1,6 +1,5 @@
 package DataTypes;
 
-
 public class Move {
     public Coordinate origin;
     public Coordinate destination;
@@ -11,17 +10,10 @@ public class Move {
         this.destination = destination;
         this.promotion = '-';
     }
-    public Move(Coordinate origin, Coordinate destination, char promotion) {
-        this.origin = origin;
-        this.destination = destination;
-        this.promotion = promotion;
-    }
 
     public Move(Move move, char promotion) {
         this.origin = move.origin;
         this.destination = move.destination;
         this.promotion = promotion;
     }
-
-    public Move() {}
 }
