+++ src/main/java/DataTypes/PieceTypes.java
@@ -1,0 +1,11 @@
+package DataTypes;
+
+public enum PieceTypes {
+    BISHOP,
+    EMPTY,
+    KING,
+    KNIGHT,
+    PAWN,
+    QUEEN,
+    ROOK
+}
