+++ src/main/java/Evaluation/EvaluationUtils.java
@@ -1,4 +1,6 @@
 package Evaluation;
 
+import DataTypes.Coordinate;
+
 public class EvaluationUtils {
 }
