+++ src/main/java/MoveGeneration/Bishop.java
@@ -1,0 +1,77 @@
+package MoveGeneration;
+
+import BoardRepresentation.BoardState;
+import DataTypes.Coordinate;
+import DataTypes.Move;
+import DataTypes.Moves;
+
+public class Bishop implements Piece {
+
+    MoveGenerationUtils moveUtils = new MoveGenerationUtils();
+    public Moves getMoves(Coordinate origin, BoardState boardState) {
+
+        char [][] board = boardState.board;
+
+        Moves moves = new Moves();
+        char orig = moveUtils.getCoordinate(origin, board);
+        int row = origin.row;
+        int col = origin.column;
+
+        //Bishop moves down and right
+        for(int i = 1; row + i <= 7 && col + i <= 7; i++) { //Runs as long as destination is within board limits
+            Coordinate destination = new Coordinate(col + i, row + i);
+            if(moveUtils.isOpposingColourOrEmpty(destination, orig, board)) {
+                moves.add(new Move(origin, destination));
+
+                if(!moveUtils.locationIsEmpty(destination, board)) {
+                    break;
+                }
+            } else {
+                break;
+            }
+        }
+
+        //Bishop moves down and left
+        for(int i = 1; row + i <= 7 && col - i >= 0; i++) {
+            Coordinate destination = new Coordinate(col - i, row + i);
+            if(moveUtils.isOpposingColourOrEmpty(destination, orig, board)) {
+                moves.add(new Move(origin, destination));
+
+                if(!moveUtils.locationIsEmpty(destination, board)) {
+                    break;
+                }
+            } else {
+                break;
+            }
+        }
+
+        //Bishop moves up and right
+        for(int i = 1; row - i >= 0 && col + i <= 7; i++) {
+            Coordinate destination = new Coordinate(col + i, row - i);
+            if(moveUtils.isOpposingColourOrEmpty(destination, orig, board)) {
+                moves.add(new Move(origin, destination));
+
+                if(!moveUtils.locationIsEmpty(destination, board)) {
+                    break;
+                }
+            } else {
+                break;
+            }
+        }
+
+        //Bishop moves up and left
+        for(int i = 1; row - i >= 0 && col - i >= 0; i++) {
+            Coordinate destination = new Coordinate(col - i, row - i);
+            if(moveUtils.isOpposingColourOrEmpty(destination, orig, board)) {
+                moves.add(new Move(origin, destination));
+
+                if(!moveUtils.locationIsEmpty(destination, board)) {
+                    break;
+                }
+            } else {
+                break;
+            }
+        }
+        return moves;
+    }
+}
