+++ src/main/java/MoveGeneration/Piece.java
@@ -1,0 +1,9 @@
+package MoveGeneration;
+
+import BoardRepresentation.BoardState;
+import DataTypes.Coordinate;
+import DataTypes.Moves;
+
+public interface Piece {
+    Moves getMoves(Coordinate origin, BoardState boardState);
+}
