+++ src/test/java/UCI_Test.java
@@ -4,6 +4,7 @@ import Main.UCI_Controller;
 import org.junit.jupiter.api.*;
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class UCI_Test {
+
     UCI_Controller uci = new UCI_Controller();
 
     @Test
